created separate config file for merge-on-main-execution and updated docs-generation config file

+++ dbt_packages/elementary/macros/utils/table_operations/delete_and_insert.sql
@@ -61,14 +61,15 @@
         {% set delete_query %}
             alter table {{ relation }} delete where
             {{ delete_column_key }} is null
-            or {{ delete_column_key }} in (select {{ delete_column_key }} from {{ delete_relation }});
+            or {{ delete_column_key }} in (select {{ delete_column_key }} from {{ delete_relation }})
+            {{ adapter.get_model_query_settings(model) }};
         {% endset %}
         {% do queries.append(delete_query) %}
     {% endif %}
 
     {% if insert_relation %}
         {% set insert_query %}
-            insert into {{ relation }} select * from {{ insert_relation }};
+            insert into {{ relation }} {{ adapter.get_model_query_settings(model) }} select * from {{ insert_relation }};
         {% endset %}
         {% do queries.append(insert_query) %}
     {% endif %}
@@ -90,6 +91,29 @@
         {% do queries.append(delete_query) %}
 
     {% elif delete_relation %}
+        {% set delete_query %}
+            delete from {{ relation }}
+            where
+            {{ delete_column_key }} is null
+            or {{ delete_column_key }} in (select {{ delete_column_key }} from {{ delete_relation }});
+        {% endset %}
+        {% do queries.append(delete_query) %}
+    {% endif %}
+
+    {% if insert_relation %}
+        {% set insert_query %}
+            insert into {{ relation }} select * from {{ insert_relation }};
+        {% endset %}
+        {% do queries.append(insert_query) %}
+    {% endif %}
+
+    {% do return(queries) %}
+{% endmacro %}
+
+{% macro redshift__get_delete_and_insert_queries(relation, insert_relation, delete_relation, delete_column_key) %}
+    {% set queries = [] %}
+
+    {% if delete_relation %}
         {% set delete_query %}
             delete from {{ relation }}
             where
